+++ Rentoza/src/main/java/org/example/rentoza/car/dto/CarRequestDTO.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 import org.example.rentoza.car.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -12,7 +13,7 @@ public class CarRequestDTO {
     private String brand;
     private String model;
     private Integer year;
-    private Double pricePerDay;
+    private BigDecimal pricePerDay;
     private String location;
     private String imageUrl;
     private List<String> imageUrls;
