+++ Rentoza/src/main/java/org/example/rentoza/owner/dto/OwnerStatsDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * DTO for owner dashboard statistics
  */
@@ -15,6 +17,6 @@ import lombok.Setter;
 public class OwnerStatsDTO {
     private Integer totalCars;
     private Integer totalBookings;
-    private Double monthlyEarnings;
+    private BigDecimal monthlyEarnings;
     private Double averageRating;
 }
